Add simple role to synchronize repositories
This role defines a systemd service/timer that synchronizes all the
repositories that I like reading using Gickup.

The system is setup to pull the Gickup configuration files from my
GitHub repository everytime it is executed, thus allowing me to
update the repository list without having to execute Ansible every
time.

Currently, the synchronization happens at 02H00 every morning, which
is the time that the internet is not used as much.

Also, everything is hard-coded for now :) will come around later
and do a sweep and get everything configured using Ansible variables.

Signed-off-by: Wynand Pieterse <wpieterse2825@gmail.com>