Allow usage on desktop machines
I have seperated out the playbook for servers and desktops so that
I can provision my local machines as well.

A script is created for servers and desktops, `run_server.sh` and
`run_desktop.sh` respectively. The script for desktop gets the local
hostname and uses that to call out to the playbook. This allows
differing configuration for whatever desktop machine there is.

For now, only the utilities-* roles are installed on desktop devices
allowing me to use consistent tools over them all.

The server playbook is kept as is for now.

Signed-off-by: Wynand Pieterse <wpieterse2825@gmail.com>